requirements.txt
requests
python-dotenv
beautifulsoup4
firecrawl
python-dotenv
exa-py
playwright
beautifulsoup4